requirements_freeze.txt
absl-py==0.9.0
alabaster==0.7.12
apex==0.1
APScheduler==3.5.3
asn1crypto==0.24.0
astor==0.8.1
atomicwrites==1.3.0
attrs==19.1.0
audioread==2.1.6
Babel==2.6.0
backcall==0.1.0
beautifulsoup4==4.7.1
bleach==3.1.0
boto3==1.9.20
botocore==1.12.213
certifi==2019.6.16
cffi==1.12.2
chardet==3.0.4
Click==7.0
codecov==2.0.15
conda==4.6.8
conda-build==3.17.8
coverage==4.5.3
cryptography==2.6.1
cycler==0.10.0
cymem==2.0.2
Cython==0.28.4
cytoolz==0.9.0.1
dataclasses==0.7
decorator==4.4.0
defusedxml==0.5.0
dill==0.2.9
docutils==0.15.2
dominate==2.5.1
entrypoints==0.3
filelock==3.0.10
flake8==3.7.7
Flask==1.0.2
future==0.17.1
gast==0.3.3
gevent==1.4.0
glob2==0.6
greenlet==0.4.15
grpcio==1.27.2
gunicorn==19.9.0
h5py==2.10.0
hypothesis==4.14.2
idna==2.7
imageio==2.8.0
imagesize==1.1.0
imgaug==0.4.0
inflect==2.1.0
ipykernel==5.1.0
ipython==7.4.0
ipython-genutils==0.2.0
itsdangerous==1.1.0
jedi==0.13.3
Jinja2==2.10
jmespath==0.9.4
joblib==0.13.2
jsonschema==3.0.1
jupyter-client==5.2.4
jupyter-core==4.4.0
jupyterlab==0.35.4
jupyterlab-server==0.2.0
jupyterlab-zip==1.0.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
libarchive-c==2.8
librosa==0.6.3
lief==0.9.0
llvmlite==0.28.0
lmdb==0.94
Markdown==3.2.1
MarkupSafe==1.1.1
maskrcnn-benchmark==0.1
matplotlib==3.0.0
mccabe==0.6.1
mistune==0.8.4
mlperf-compliance==0.0.10
mock==2.0.0
more-itertools==7.0.0
msgpack==0.6.1
msgpack-numpy==0.4.3.2
murmurhash==1.0.2
nbconvert==5.4.1
nbformat==4.4.0
networkx==2.0
newrelic==4.18.0.118
ninja==1.9.0
nltk==3.4
nmslib==2.0.6
nose==1.3.7
notebook==5.7.5
numba==0.43.1
numpy==1.17.0
nvidia-dali==0.8.1
onnx==1.4.1
opencv-python==3.4.1.15
packaging==19.0
pandas==0.24.2
pandocfilters==1.4.2
parso==0.3.4
pbr==5.1.3
pexpect==4.6.0
pickleshare==0.7.5
Pillow==5.3.0
Pillow-SIMD==5.3.0.post0
pkginfo==1.5.0.1
plac==0.9.6
pluggy==0.9.0
preshed==2.0.1
prometheus-client==0.6.0
prompt-toolkit==2.0.9
protobuf==3.11.3
psutil==5.6.1
ptyprocess==0.6.0
py==1.8.0
pybind11==2.2.4
pycocotools==2.0+nv0.2.1
pycodestyle==2.5.0
pycosat==0.6.3
pycparser==2.19
pydot==1.4.1
pyflakes==2.1.1
Pygments==2.3.1
pyOpenSSL==19.0.0
pyparsing==2.4.2
pypng==0.0.20
pyrsistent==0.14.11
PySocks==1.6.8
pytest==4.4.0
pytest-cov==2.6.1
pytest-pythonpath==0.7.3
python-dateutil==2.8.0
python-nvd3==0.15.0
python-slugify==3.0.2
pytz==2019.2
PyWavelets==1.1.1
PyYAML==5.1
pyzmq==18.0.1
regex==2018.1.10
requests==2.19.1
resampy==0.2.1
revtok==0.0.3
ruamel-yaml==0.15.46
s3transfer==0.1.13
sacrebleu==1.2.10
sacremoses==0.0.13
scikit-image==0.16.2
scikit-learn==0.20.3
scipy==1.2.1
Send2Trash==1.5.0
sensei-sdk==2.2
senseinbs==2.13
sentencepiece==0.1.92
Shapely==1.7.0
singledispatch==3.4.0.3
six==1.12.0
snowballstemmer==1.2.1
soupsieve==1.8
spacy==2.0.16
Sphinx==2.0.0
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.1
SSD==0.1
tabulate==0.8.3
tensorboard==1.10.0
termcolor==1.1.0
terminado==0.8.2
testpath==0.4.2
text-unidecode==1.2
thinc==6.12.1
tokenizers==0.7.0
toolz==0.9.0
torch==1.1.0a0+9eb0f43
torchsummary==1.5.1
torchtext==0.4.0
torchvision==0.2.1
tornado==5.1.1
tqdm==4.31.1
traitlets==4.3.2
transformers==2.11.0
typing==3.6.6
typing-extensions==3.7.2
tzlocal==2.0.0
ujson==1.35
Unidecode==1.0.23
urllib3==1.23
wcwidth==0.1.7
webencodings==0.5.1
Werkzeug==0.15.1
wrapt==1.10.11
wstunnel==0.0.6
yacs==0.1.6
